add get and set app id, get and set one signal id to Operation Class
