+++ src/core/operations/Operation.ts
@@ -1,3 +1,4 @@
+import { IDManager } from 'src/shared/managers/IDManager';
 import { Model } from '../models/Model';
 
 export const GroupComparisonType = {
@@ -10,30 +11,54 @@ export type GroupComparisonValue =
   (typeof GroupComparisonType)[keyof typeof GroupComparisonType];
 
 export abstract class Operation extends Model {
-  constructor(name: string) {
+  constructor(name: string, appId?: string, onesignalId?: string) {
     super();
     this.name = name;
+    if (appId && onesignalId) {
+      this.appId = appId;
+      this.onesignalId = onesignalId;
+    }
   }
 
   get name(): string {
     return this.getProperty<string>('name');
   }
-
   private set name(value: string) {
     this.setProperty<string>('name', value);
+  }
+
+  /**
+   * The application ID this subscription will be created under.
+   */
+  get appId(): string {
+    return this.getProperty<string>('appId');
+  }
+  protected set appId(value: string) {
+    this.setProperty<string>('appId', value);
+  }
+
+  get onesignalId(): string {
+    return this.getProperty<string>('onesignalId');
+  }
+  protected set onesignalId(value: string) {
+    this.setProperty<string>('onesignalId', value);
   }
 
   /**
    * This is a unique id that points to a record this operation will affect.
    * Example: If the operation is updating tags on a User this will be the onesignalId.
    */
-  abstract get applyToRecordId(): string;
+  get applyToRecordId(): string {
+    return this.onesignalId;
+  }
 
   /**
    * The key of this operation for when the starting operation has a `groupComparisonType`
    * of `GroupComparisonType.CREATE`
    */
-  abstract get createComparisonKey(): string;
+  get createComparisonKey(): string {
+    return '';
+  }
 
   /**
    * The key of this operation for when the starting operation has a `groupComparisonType`
@@ -50,14 +75,18 @@ export abstract class Operation extends Model {
   /**
    * Whether the operation can currently execute given its current state.
    */
-  abstract get canStartExecute(): boolean;
+  get canStartExecute(): boolean {
+    return !IDManager.isLocalId(this.onesignalId);
+  }
 
   /**
    * Called when an operation has resolved a local ID to a backend ID.
    * Any IDs within the operation that could be local should be translated at this time.
    */
   translateIds(map: Record<string, string>): void {
-    // Optional override
+    if (map[this.onesignalId]) {
+      this.onesignalId = map[this.onesignalId];
+    }
   }
 
   toString(): string {
